requirements.txt
flask>=3.0,<4.0
pydantic>=2.0,<3.0
openai>=1.0,<2.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
pytest>=8.0,<9.0
pytest-mock>=3.12,<4.0
